azure-functions/requirements.txt
azure-functions
python-dotenv
pillow
# Add any other dependencies your vision_ocr.py needs